use 32bit types where possible, remove system call to sort

+++ tests/test_core.py
@@ -93,11 +93,11 @@ def testProcessNoiseAdjustment():
     dStatPC = 1.0
     inflatedQ = False
 
-    matrixQ = np.array([[minQ, offDiag], [offDiag, minQ]])
+    matrixQ = np.array([[minQ, offDiag], [offDiag, minQ]], dtype=np.float32)
     matrixQCopy = matrixQ.copy()
-    vectorY = (np.random.normal(0, 15, size=m))
-    dStat = np.mean(vectorY**2)
-    dStatDiff = np.sqrt(np.abs(dStat - dStatAlpha)*dStatd + dStatPC)
+    vectorY = (np.random.normal(0, 15, size=m)).astype(np.float32)
+    dStat = np.mean(vectorY**2).astype(np.float32)
+    dStatDiff = np.float32(np.sqrt(np.abs(dStat - dStatAlpha)*dStatd + dStatPC))
 
     matrixQ, inflatedQ = cconsenrich.updateProcessNoiseCovariance(matrixQ,
     matrixQCopy,
